+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Features
 
+- Add Staking routes;
+
 ### Improvements
 
 ### Improvements
@@ -11,9 +13,11 @@
 # v0.5.2 (2019.02.25)
 
 ### Features
+
 - add CENNZScan address link to account public address
 
 ### Improvements
+
 - tidy up asset table layout
 - update background gradient settings
 - update field text bg color to match with design