+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features
 - able to get seed from wallet json
 - able to stake and unstake
+- Add Staking routes;
 
 ### Fixes
 
@@ -17,9 +18,11 @@
 # v0.5.2 (2019.02.25)
 
 ### Features
+
 - add CENNZScan address link to account public address
 
 ### Improvements
+
 - tidy up asset table layout
 - update background gradient settings
 - update field text bg color to match with design